+++ R/config.R
@@ -12,10 +12,10 @@
 #'
 #' @examples
 #' \dontrun{
-#' foundry_set_endpoint("AZURE_FOUNDRY_ENDPOINT")
+#' foundry_set_endpoint(Sys.getenv("AZURE_FOUNDRY_ENDPOINT"))
 #'
 #' # Store permanently
-#' foundry_set_endpoint("AZURE_FOUNDRY_ENDPOINT", store = TRUE)
+#' foundry_set_endpoint(Sys.getenv("AZURE_FOUNDRY_ENDPOINT"), store = TRUE)
 #' }
 foundry_set_endpoint <- function(endpoint, store = FALSE) {